+++ packages/nominal-smartcard/tests/test_cert_selection.py
@@ -1,10 +1,10 @@
 from __future__ import annotations
 
 import pytest
+from _helpers import _candidate, _make_der_cert
 
 from nominal.smartcard._cert_selection import _assert_client_auth_eku, select_piv_authentication_certificate
 from nominal.smartcard._errors import SmartcardCertificateSelectionError
-from tests.smartcard._helpers import _candidate, _make_der_cert
 
 # CertificateCandidate.is_piv_authentication_candidate
