+++ packages/nominal-smartcard/tests/test_transport.py
@@ -5,11 +5,11 @@ import threading
 from pathlib import Path
 
 import pytest
+from _helpers import _candidate, _FakeBackend, _make_der_cert
 
 from nominal.smartcard._pkcs11 import NOMINAL_PKCS11_MODULE_ENV_VAR
 from nominal.smartcard._session import SmartcardSession, SmartcardSessionManager
 from nominal.smartcard._transport import SmartcardSslContextProvider
-from tests.smartcard._helpers import _candidate, _FakeBackend, _make_der_cert
 
 
 class _FakeBridge:
